fix: handle async generic arrow functions (#20)
Closes #19

This updates the transform to detect `async` arrow functions that have
type parameters which span over multiple lines. When this is detected
the `(` at the start of the functions parameters is moved to the start
of the type parameters. Avoiding a new line character to appear directly
after the `async` keyword.

Signed-off-by: Ashley Claymore <aclaymore@bloomberg.net>