fix(playground): support sharing unicode input (#37)
closes #34

This PR updates the base64 encode/decode logic to be done in utf8,
previously it was going straight to `btoa` which only supports ASCII.
This also means that existing links created will be forwards compatible
with the new decoding.

**Testing performed**

`cd website; npm start`

Creating playgrounds with unicode input were sharable. I also took an
existing url from the currently deployed playground and checked that the
new logic is backwards compatible with existing links.

Signed-off-by: Ashley Claymore <aclaymore@bloomberg.net>